refactor(email): moved to /templates directory
Moving email-templates into the templates directory as part of creating
all of the /layout content. Also necessitated generating an index.html
file, and updated the site.navigation variable. See slack message on
6/15 in team channel.

—

Evan Nabors [1:49 PM]
Where should this page move to helix.rax.io/layout/email-template.html?
In the I.A. we have moved it to a templates top level item that
probably won’t be ready for some time, but it no longer seems to fit
into `/layout`.  So let’s take a vote, should we 1. keep it in layout
section 2. move it to components for the time being or 3. create the
templates item with this being the sole item under it for the time
being.