feature(Anchor Tags): Create Dynamic h2 Anchors
Built some small js that parses the page for H2 elements and adds them
to an anchor-list ul, if present on the page. Right now this element
must be declared on the page but we will want to dynamically create it
in the future. Just unsure where to put it so I have omitted this for
the time being. We also need to create the styles for the UL, which I
can’t seem to get working. Take a look at line 370 of _layout.scss to
see if you can figure it out.