feature(lightbox): Add libraries
Lightbox libraries, icons, styles checked in and included in the
necessary locations for templates. Feature has been tested but will not
be turned on for any images. To use see the source docs.
http://lokeshdhakar.com/projects/lightbox2/. Also Laura may be
including a jekyll plugin (ruby library) that converts markdown into
the format for this library to make it easier for content contributors.