Fix family pages (#278)
* refactor(scss): removes article element rule, adds content-wrapper

* refactor(scss): removes padding on banner and article elements

* refactor(scss): removes padding on title and removes nesting on component-row

* feat(templates): creates component row file for family pages

* refactor(scss): removes padding and converts article rule to content-wrapper class

* refactor(html): adds content-wrapper class

* feat(html): adds content-wrapper class

This facilitates centering and max-width of content

* feat(html): adds content-wrapper class

* refactor(html): adds content-wrapper class and applies hx columns to entire header

* refactor(html): replaces raw html with included component-row.html

* refactor(scss): removes horizontal margin on com-code rule

* refactor(html): moves com-code inside content-wrapper

* refactor(scss): removes redundant content-wrapper rule and anchor styling

I removed the anchor styling on the toc so that the link colors would match the styles in Ty’s illustrator docs.

* refactor(scss): removes redundant content-wrapper and toc link rules

This also removes the H* styles.  As I toggled between deleting the H* styles and keeping them, I saw no difference in the page.

Deleting for now.  As I look at the new docs MD files for Tabs, etc.  I think we need to adjust the contribution template to use H* markdown slightly differently than what’s currently specified.

* refactor(scss): removes nesting around the markdown-toc

removed the

* refactor(scss): removes nesting around the markdown-toc rule