Add two column layout to component details (#281)
* refactor(SVG): removes backgrounds on tabs SVGs

* refactor(md file): renames tabs.md to tabset.md

* refactor(scss): sets body background to #fff with override

* refactor(scss): sets overview-wrapper background to #fafafa per Ty’s suggestion

* refactor(includes): updates toc to target H3s & H4s

* refactor(scss): removes commented code

The H2 rule here was messing with H2 rules from Helix.  Removed it.

I added the base line height to the P tag and remove top margin from them as well.

* refactor(scss): removes base element rules

This is an effort to start moving toward just Helix styles in the website.

* feat(scss): adds rule to remove top margin and reduce bottom mardin on H4s in component-content

Let’s figure out what the best practice is here.  The sizing and coloring of H* tags coming from helix-ui.css should be the rule, but is it ok to tinker with margins to get the website to design spec?

* refactor(markdown): adds two-column responsive layout

This changes H2s to H3s, and uses H4s for the subsection headers since they will ultimately match Ty’s design more closely when helix is updated with accurate font-specs.

* refactor(scss): removes forms import from website.scss

Not sure its needed at all anymore, and it was jacking with styling on the component page.

* refactor(SVG): updates from sketch docs

* feat(markdown): adds responsive classes to page-layouts

* refactor(markdown): change from Designing with tabsets to Design Specs

* refactor(markdown): adds columns with responsive classes

* refactor(markdown): adds helix columns and responsive classes

* refactor(markdown): updates alert-bar usage description

* refactor(markdown): updates best practice description

* refactor(markdown): clarifies text wrapping behavior notes

* refactor(markdown): updates status types description per feedback