Preview icons usage/evan (#286)
* docs(nav): capitalize home link

Bandaid capitalize HOME in writing until we look more in-depth into
writing a custom class to uppercase.

* refactor(content container): remove padding

Overwrite padding pulled in from hx—app for docs site alone.

* refactor(family-page): spacing on blurb

Removed margins and increase line-height to match redline spec.

* refactor(doc page): adjust content size of component blurb

* docs(component-usage): Write usage info for most components and include new preview images for 'content areas'

* update(modal): provide preview-image from CDN

* feature(preview-icons): built new preview icons for all existing components.

Everything that is currently displayed on Family pages now has a nifty
new SVG in the style given in the redlines file, sourced from CDN for
the time being.

* fix(figure spacing): Small tweak to add additional padding to figcaption

Currently when you use two figures vertically with a column the spacing
between the two figures and the fig caption between them is not enough
to discern wither a caption applies to the figure above or below. This
tweak adds additionally padding to assist in seeing this.

* fix(requested changes): Small tweaks based on last review

This updates the variable used for component blurbs as well as updates
the usage data with updated verbiage from Bart.

* fix(editorial review): Tiny editorial changes requested

Addressing tiny word choice level changes based on feedback from Laura
on this PR.