Thumbnail update/evan (#290)
* docs(footer): Small formatting updates

This PR fixes the items noted in the following image from PR 283.
https://user-images.githubusercontent.com/933210/31347430-fadbc5dc-ace1-
11e7-9b92-e0622c4609ad.png.  Additionally a 1rem bottom-margin was
added to H3 elements that are wrapped inside the `component-content`
div elements so this change is scoped to only Component page content.

* edit(svg): footer preview SVG

Edits description, remove nesting level, and rename id's for
preview-image. None of these changes effect the graphic output.
Similarly none of the ID’s are being used, this is just for the sake of
the cleanliness of the SVG itself.