fix(revert-h3-margins): Reverting spacing on h3
This went in as part of PR-290 and a better means to accomplish this
has been found by Laura.

———

Evan Nabors [9:56 AM]
Does that ^ mean I should revert the 1rem margin-bottom on h3's in the
footer PR and use this process instead?

[9:56]
https://github.com/rackerlabs/design-system/pull/290/files#diff-1709761f
33d26e33b1c43086716758f5 (edited)

[9:56]
change in question

Laura Santamaria [9:57 AM]
excellent! and @evan.nabors, yeah, probably. That way, we don’t have to
worry about it long-term should the design get adjusted