feat(scss): adds height 100% to .image elements so the figure fills its vertical space
Note that if this makes things look funky, we can revert this.

I changed this because of the fact that in rows where we have more than one figure image, the container for the shorter image didn’t reach the bottom of the container for the taller image, and it looked funky.