refactor(markdown): removes kramdown for title classes
After talking with Laura, we decided to use normal # notation for headers in the markdown, starting with H2s and H3s since the layout header contains the H1.

The plan is to use a plugin to auto-apply correct Helix classes so that we don’t end up with various pages featuring different types of headers.