Fix formatting of footer website page surf 655 (#292)
* refactor(footer): consolidates information and applies HX header classes

* refactor(scss): adjusts the way spacing is handled in component banner

Removes the rules for the hero image as the thumbnail sizing is handled by the HX responsive column classes.
Adds padding to the content-wrapper for a more simplified way to handle spacing in the banner

* refactor(footer): switches to proper hierarchy in the # headers

* refactor(template): consolidates html in the banner

* bug(markdown): fixes typoes in application-navigation and breadcrumbs

* feat(scss): changes component page blurb to match Ty’s design

* refactor(includes): changes TOC back to targeting H2s and H3s

* refactor(markdown): removes kramdown for title classes

After talking with Laura, we decided to use normal # notation for headers in the markdown, starting with H2s and H3s since the layout header contains the H1.

The plan is to use a plugin to auto-apply correct Helix classes so that we don’t end up with various pages featuring different types of headers.

* refactor(markdown): updates footer.md per Laura’s notes