docs(accessibility): segment out page (#307)
* docs(accessibility): segment out page

first commit to build out accessibility page and built out SVGs and
/assets/ folder structure.

* fix(svgs): SVGs were created using png's, retraced in vector format

I did not have the source SVGs so I recreated them using sketch.

* docs(meta-data): remove unnecessary meta-data

Removed the Parent meta-data to ensure the page will not appear in the
incorrect location. Removed TOC from doc, removed preview-image as it
did not apply to this doc. Edited titles

* fix(navigation): update navigation, add principles, move files

The accessibility.md file was placed into a component directory as a
first pass for editing and displaying. This commit moves the file to
the correct location and updates navigation to include /principles as
the second item.

* docs(accessibility): updates post editorial review

* docs(accessibility): one sentence edit

* docs(accessibility): add usage text

* docs(accessibility): content edits post review

* svgs(accessibility): add `@include` google fonts to images; fix inputs-2 image

* docs(accessibility): fix typo