docs(getting-started): convert GSG sketch to md.
Adding in getting started content. Linked to from homepage. Also added
new helper classes for hxRow to add padding-bottom. Only used in the
Getting started doc. This can be removed, or modified based on
feedback. Also, the homepage button link was updated to point to the
GSG, an extension for hxBtn -> a for the cards was annotated and added
to the _home.scss file as well.