docs(getting-started): convert GSG sketch to md. (#327)
* docs(getting-started): convert GSG sketch to md.

Adding in getting started content. Linked to from homepage. Also added
new helper classes for hxRow to add padding-bottom. Only used in the
Getting started doc. This can be removed, or modified based on
feedback. Also, the homepage button link was updated to point to the
GSG, an extension for hxBtn -> a for the cards was annotated and added
to the _home.scss file as well.

* docs(GSG): update with newest content

* docs(getting-started): move/rename file, update nav, include edits from PR review.

Had to move the file to create a prettier URL.

* docs(get-start): update css / md based on review

* docs(get-started): updated title

* docs(get-started): liquid template remove '-' in title, update page name