Docs status labels (#401)
* fix(tabs-preview-image): resize SVG

This svg was not correctly scaling and didn’t line up with the other
images. New image created to fix this.

* fix(search-preview image): remove opacity rules

This image was using #F57C00 with an opacity applied. This was not
intentional and is the only image that used this style. This was
ditched for consistency.

* docs(status-labels): add status label docs

converted the status-indicators.md file to status-labels.md and add
content.

* docs(status-labels): add 2x-pngs and svgs

* docs(status-labels): include content from bart, rename title

* docs(status): first content edits

* docs(status): edits for bart

* docs(status): edits for laura

* docs(status): updates for laura 2.0

* docs(status): sentence edit for bart

* docs(labels): updates for drew

* docs(status): remove references to health status labels

* docs(status): one sentence edit per bowling review